# Conflicts:
#	group-investment/src/main/java/com/example/group_investment/member/Member.java
#	group-investment/src/main/java/com/example/group_investment/team/TeamService.java
#	stock-system/src/main/java/com/example/stock_system/account/Account.java
#	stock-system/src/main/java/com/example/stock_system/account/TeamAccountRepository.java
#	stock-system/src/main/resources/application.yml

+++ chatting/src/main/java/com/example/chatting/repository/ChatRoomRepository.java
@@ -3,6 +3,6 @@ package com.example.chatting.repository;
 import com.example.chatting.domain.ChatRoom;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ChatRoomRepository extends CrudRepository<ChatRoom, Long> {
+public interface ChatRoomRepository extends CrudRepository<ChatRoom, Integer> {
     //ChatRoom findChatRoomById(Long id);
 }
